requirements.txt
APScheduler==3.6.3
beautifulsoup4==4.10.0
cachetools==4.2.2
certifi==2021.5.30
charset-normalizer==2.0.4
cycler==0.10.0
et-xmlfile==1.1.0
idna==3.2
kiwisolver==1.3.2
lxml==4.6.3
matplotlib==3.4.3
mplfinance==0.12.7a17
multitasking==0.0.9
numpy==1.21.2
pandas==1.3.3
pandas-datareader==0.10.0
Pillow==8.3.2
pyparsing==2.4.7
python-dateutil==2.8.2
python-telegram-bot==13.10
pytz==2021.1
pytz-deprecation-shim==0.1.0.post0
requests==2.26.0
six==1.16.0
soupsieve==2.2.1
tornado==6.1
tzdata==2021.5
tzlocal==4.1
urllib3==1.26.6
xlrd==2.0.1
yfinance==0.1.63
